+++ audioquake/launcherlib/ui/tabs/play.py
@@ -1,0 +1,56 @@
+"""AudioQuake Game Launcher - Play tab"""
+from os import path
+
+import wx
+
+from launcherlib.ui.helpers import \
+	add_launch_button, add_opener_buttons, add_widget
+from launcherlib.game_controller import on_windows
+
+
+class PlayTab(wx.Panel):
+	def __init__(self, parent, game_controller):
+		wx.Panel.__init__(self, parent)
+		sizer = wx.BoxSizer(wx.VERTICAL)
+
+		# Launching the game
+
+		play = wx.StaticBoxSizer(wx.VERTICAL, self, "Play")
+
+		game_modes = {
+			"Play": game_controller.launch_default,
+			"Tutorial": game_controller.launch_tutorial,
+		}
+
+		for title, action in game_modes.items():
+			add_launch_button(self, play, title, action)
+
+		if on_windows():
+			server = ('zqds.exe',)
+			rcon = ('rcon.exe', '--ask')
+		else:
+			server = ('./start-server.command',)
+			rcon = ('./start-rcon.command',)
+
+		add_opener_buttons(self, play, {
+			"Dedicated server": server,
+			"Remote console": rcon,
+		})
+
+		add_widget(sizer, play)
+
+		# Help
+
+		docs = wx.StaticBoxSizer(wx.VERTICAL, self, "Help")
+
+		add_opener_buttons(self, docs, {
+			'User manual': (path.join('manuals', 'user-manual.html'),),
+			'Sound legend': (path.join('manuals', 'sound-legend.html'),),
+		})
+
+		add_widget(sizer, docs)
+
+		# Wiring
+
+		sizer.SetSizeHints(self)
+		self.SetSizer(sizer)
